Added command details folder (#259)
